Refactor models and services
