+++ Sources/Passes/PassesError.swift
@@ -5,8 +5,6 @@
 //  Created by Scott Grosch on 1/22/20.
 //
 
-import Foundation
-
 public enum PassesError: Error {
     /// The template path is not a directory
     case templateNotDirectory
